+++ src/handlers/motions/motionCreated/handlers/index.ts
@@ -1,0 +1,3 @@
+export { handleMintTokensMotion } from './mintTokens';
+export { handleNetworkUpgradeMotion } from './networkUpgrade';
+export { handleManageDomainMotion } from './manageDomain';
