+++ src/handlers/motions/motionCreated/helpers.ts
@@ -2,17 +2,12 @@ import { AnyColonyClient, Extension } from '@colony/colony-js';
 import { BigNumber } from 'ethers';
 import { TransactionDescription } from 'ethers/lib/utils';
 
-import { MotionData, ContractEvent, motionNameMapping } from '~types';
-import {
-  getColonyTokenAddress,
-  getDomainDatabaseId,
-  getVotingClient,
-  verbose,
-  getPendingMotionDomainDatabaseId,
-} from '~utils';
+import { MotionData, ContractEvent } from '~types';
+import { getVotingClient, verbose } from '~utils';
+
+import { mutate } from '~amplifyClient';
 
 import {
-  createMotionInDB,
   getMotionDatabaseId,
   getRequiredStake,
   getUserMinStake,
@@ -38,7 +33,7 @@ export const getParsedActionFromMotion = async (
   }
 };
 
-const getMotionData = async (
+export const getMotionData = async (
   colonyAddress: string,
   motionId: BigNumber,
   domainId: BigNumber,
@@ -100,52 +95,27 @@ const getMotionData = async (
   };
 };
 
-export const writeMintTokensMotionToDB = async (
+export const createMotionInDB = async (
   {
     transactionHash,
-    contractAddress: colonyAddress,
     blockNumber,
+    contractAddress: colonyAddress,
     args: { motionId, creator, domainId },
   }: ContractEvent,
-  parsedAction: TransactionDescription,
+  input: Record<string, any>,
 ): Promise<void> => {
-  const { name, args: actionArgs } = parsedAction;
-  const amount = actionArgs[0].toString();
-  const tokenAddress = await getColonyTokenAddress(colonyAddress);
-  const motionData = await getMotionData(colonyAddress, motionId, domainId);
-  await createMotionInDB({
-    id: transactionHash,
-    colonyId: colonyAddress,
-    type: motionNameMapping[name],
-    motionData,
-    tokenAddress,
-    fromDomainId: getDomainDatabaseId(colonyAddress, domainId),
-    initiatorAddress: creator,
-    amount,
-    blockNumber,
-  });
-};
-
-export const writeManageDomainMotionToDB = async (
-  {
-    transactionHash,
-    contractAddress: colonyAddress,
-    blockNumber,
-    args: { motionId, creator, domainId },
-  }: ContractEvent,
-  parsedAction: TransactionDescription,
-): Promise<void> => {
-  const { name } = parsedAction;
-  const pendingDomainMetadataId = getPendingMotionDomainDatabaseId(colonyAddress, transactionHash);
   const motionData = await getMotionData(colonyAddress, motionId, domainId);
 
-  await createMotionInDB({
-    id: transactionHash,
-    colonyId: colonyAddress,
-    type: motionNameMapping[name],
-    motionData,
-    initiatorAddress: creator,
-    blockNumber,
-    pendingDomainMetadataId,
+  await mutate('createColonyAction', {
+    input: {
+      id: transactionHash,
+      colonyId: colonyAddress,
+      isMotion: true,
+      showInActionsList: false,
+      motionData,
+      initiatorAddress: creator,
+      blockNumber,
+      ...input,
+    },
   });
 };
