+++ src/handlers/motions/motionCreated/motionCreated.ts
@@ -1,11 +1,12 @@
 import networkClient from '~networkClient';
 import { ColonyOperations, ContractEvent } from '~types';
 import { verbose } from '~utils';
+import { getParsedActionFromMotion } from './helpers';
 import {
-  getParsedActionFromMotion,
-  writeManageDomainMotionToDB,
-  writeMintTokensMotionToDB,
-} from './helpers';
+  handleManageDomainMotion,
+  handleMintTokensMotion,
+  handleNetworkUpgradeMotion,
+} from './handlers';
 
 export default async (event: ContractEvent): Promise<void> => {
   const {
@@ -22,14 +23,20 @@ export default async (event: ContractEvent): Promise<void> => {
     /* Handle the action type-specific mutation here */
     switch (contractOperation) {
       case ColonyOperations.MintTokens: {
-        await writeMintTokensMotionToDB(event, parsedAction);
+        await handleMintTokensMotion(event, parsedAction);
         break;
       }
       case ColonyOperations.AddDomain:
       case ColonyOperations.EditDomain: {
-        await writeManageDomainMotionToDB(event, parsedAction);
+        await handleManageDomainMotion(event, parsedAction);
         break;
       }
+
+      case ColonyOperations.Upgrade: {
+        await handleNetworkUpgradeMotion(event, parsedAction);
+        break;
+      }
+
       default: {
         break;
       }
